+++ next.config.ts
@@ -11,6 +11,7 @@ const nextConfig: NextConfig = {
   env: {
     AIRTABLE_ACCESS_TOKEN: process.env.AIRTABLE_ACCESS_TOKEN,
     AIRTABLE_BASE_ID: process.env.AIRTABLE_BASE_ID,
+    AIRTABLE_TABLE_NAME: process.env.AIRTABLE_TABLE_NAME,
   },
   async headers() {
     return [